requirements.txt
PyGithub==2.8.1
requests==2.32.5
azure-ai-projects==1.0.0
azure-ai-agents==1.1.0
azure-core==1.38.2
azure-storage-blob==12.28.0
cryptography==46.0.5
certifi==2026.2.25
